+++ test/VoucherHub.test.ts
@@ -32,8 +32,8 @@ const initVoucherHubBalance = 10n * voucherValue; // arbitrary value, but should
 describe('VoucherHub', function () {
     let iexecPoco: string;
     let iexecPocoInstance: IexecPocoMock;
-    let voucherHubWithVoucherManagerSigner: VoucherHub;
-    let voucherHubWithAssetEligibilityManagerSigner: VoucherHub;
+    let voucherHubWithMinterSigner: VoucherHub;
+    let voucherHubWithManagerSigner: VoucherHub;
     let voucherHubWithAnyoneSigner: VoucherHub;
     let voucherHubAddress: string;
     // We define a fixture to reuse the same setup in every test.
@@ -41,14 +41,8 @@ describe('VoucherHub', function () {
     // and reset Hardhat Network to that snapshot in every test.
     async function deployFixture() {
         // Contracts are deployed using the first signer/account by default
-        const [
-            admin,
-            assetEligibilityManager,
-            voucherManager,
-            voucherOwner1,
-            voucherOwner2,
-            anyone,
-        ] = await ethers.getSigners();
+        const [admin, manager, minter, voucherOwner1, voucherOwner2, anyone] =
+            await ethers.getSigners();
         const beacon = await voucherUtils.deployBeaconAndImplementation(admin.address);
 
         iexecPocoInstance = await new IexecPocoMock__factory()
@@ -57,13 +51,13 @@ describe('VoucherHub', function () {
             .then((x) => x.waitForDeployment());
         iexecPoco = await iexecPocoInstance.getAddress();
         const voucherHub = await voucherHubUtils.deployHub(
-            assetEligibilityManager.address,
-            voucherManager.address,
+            manager.address,
+            minter.address,
             iexecPoco,
             await beacon.getAddress(),
         );
-        voucherHubWithVoucherManagerSigner = voucherHub.connect(voucherManager);
-        voucherHubWithAssetEligibilityManagerSigner = voucherHub.connect(assetEligibilityManager);
+        voucherHubWithMinterSigner = voucherHub.connect(minter);
+        voucherHubWithManagerSigner = voucherHub.connect(manager);
         voucherHubWithAnyoneSigner = voucherHub.connect(anyone);
         voucherHubAddress = await voucherHub.getAddress();
         await iexecPocoInstance
@@ -73,8 +67,8 @@ describe('VoucherHub', function () {
             beacon,
             voucherHub,
             admin,
-            assetEligibilityManager,
-            voucherManager,
+            manager,
+            minter,
             voucherOwner1,
             voucherOwner2,
             anyone,
@@ -83,29 +77,16 @@ describe('VoucherHub', function () {
 
     describe('Initialize', function () {
         it('Should initialize', async function () {
-            const { beacon, voucherHub, admin, assetEligibilityManager, voucherManager } =
-                await loadFixture(deployFixture);
+            const { beacon, voucherHub, admin, manager, minter } = await loadFixture(deployFixture);
             const voucherBeaconAddress = await beacon.getAddress();
             // Check roles.
             expect(await voucherHub.owner())
                 .to.equal(await voucherHub.defaultAdmin())
                 .to.equal(admin);
             expect(await voucherHub.defaultAdminDelay()).to.equal(0);
-            expect(
-                await voucherHub.hasRole(await voucherHub.UPGRADE_MANAGER_ROLE.staticCall(), admin),
-            );
-            expect(
-                await voucherHub.hasRole(
-                    await voucherHub.ASSET_ELIGIBILITY_MANAGER_ROLE.staticCall(),
-                    assetEligibilityManager,
-                ),
-            );
-            expect(
-                await voucherHub.hasRole(
-                    await voucherHub.VOUCHER_MANAGER_ROLE.staticCall(),
-                    voucherManager,
-                ),
-            );
+            expect(await voucherHub.hasRole(await voucherHub.UPGRADER_ROLE.staticCall(), admin));
+            expect(await voucherHub.hasRole(await voucherHub.MANAGER_ROLE.staticCall(), manager));
+            expect(await voucherHub.hasRole(await voucherHub.MINTER_ROLE.staticCall(), minter));
             // Check config.
             expect(await voucherHub.getIexecPoco()).to.equal(iexecPoco);
             expect(await voucherHub.getVoucherBeacon()).to.equal(voucherBeaconAddress);
@@ -118,16 +99,10 @@ describe('VoucherHub', function () {
         });
 
         it('Should not initialize twice', async function () {
-            const { beacon, voucherHub, assetEligibilityManager, voucherManager } =
-                await loadFixture(deployFixture);
+            const { beacon, voucherHub, manager, minter } = await loadFixture(deployFixture);
 
             await expect(
-                voucherHub.initialize(
-                    assetEligibilityManager,
-                    voucherManager,
-                    iexecPoco,
-                    await beacon.getAddress(),
-                ),
+                voucherHub.initialize(manager, minter, iexecPoco, await beacon.getAddress()),
             ).to.be.revertedWithCustomError(voucherHub, 'InvalidInitialization');
         });
     });
@@ -161,11 +136,10 @@ describe('VoucherHub', function () {
     describe('Create voucher type', function () {
         it('Should create a voucher type when the caller is authorized', async function () {
             const { voucherHub } = await loadFixture(deployFixture);
-            const createTypeTx =
-                await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                    description,
-                    duration,
-                );
+            const createTypeTx = await voucherHubWithManagerSigner.createVoucherType(
+                description,
+                duration,
+            );
             const type0 = await voucherHub.getVoucherType(0);
             const count = await voucherHub.getVoucherTypeCount();
             // Run assertions.
@@ -197,16 +171,10 @@ describe('VoucherHub', function () {
     describe('Update voucher type description', function () {
         const newDescription = 'Long Term Duration';
         it('Should update voucher description', async function () {
-            const { voucherHub, assetEligibilityManager } = await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             const updateDescriptionTx =
-                await voucherHubWithAssetEligibilityManagerSigner.updateVoucherTypeDescription(
-                    0,
-                    newDescription,
-                );
+                await voucherHubWithManagerSigner.updateVoucherTypeDescription(0, newDescription);
             await updateDescriptionTx.wait();
             expect(updateDescriptionTx)
                 .to.emit(voucherHub, 'VoucherTypeDescriptionUpdated')
@@ -214,28 +182,18 @@ describe('VoucherHub', function () {
         });
 
         it('Should not update voucher description when the caller is not authorized', async function () {
-            const { voucherHub, assetEligibilityManager, anyone } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             await expect(
                 voucherHubWithAnyoneSigner.updateVoucherTypeDescription(0, newDescription),
             ).to.be.revertedWithCustomError(voucherHub, 'AccessControlUnauthorizedAccount');
         });
 
         it('Should not update description when the voucher type ID is out of bounds', async function () {
-            const { voucherHub, assetEligibilityManager } = await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             await expect(
-                voucherHubWithAssetEligibilityManagerSigner.updateVoucherTypeDescription(
-                    999,
-                    newDescription,
-                ),
+                voucherHubWithManagerSigner.updateVoucherTypeDescription(999, newDescription),
             ).to.be.revertedWith('VoucherHub: type index out of bounds');
         });
     });
@@ -243,16 +201,12 @@ describe('VoucherHub', function () {
     describe('Update voucher type duration', function () {
         const newDuration = 7200;
         it('Should update voucher duration', async function () {
-            const { voucherHub, assetEligibilityManager } = await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
+            const { voucherHub } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
+            const updateDurationTx = await voucherHubWithManagerSigner.updateVoucherTypeDuration(
+                0,
+                newDuration,
             );
-            const updateDurationTx =
-                await voucherHubWithAssetEligibilityManagerSigner.updateVoucherTypeDuration(
-                    0,
-                    newDuration,
-                );
             await updateDurationTx.wait();
             expect(updateDurationTx)
                 .to.emit(voucherHub, 'VoucherTypeDurationUpdated')
@@ -260,51 +214,39 @@ describe('VoucherHub', function () {
         });
 
         it('Should not update voucher duration when the caller is not authorized', async function () {
-            const { voucherHub, assetEligibilityManager, anyone } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             await expect(
                 voucherHubWithAnyoneSigner.updateVoucherTypeDuration(0, newDuration),
             ).to.be.revertedWithCustomError(voucherHub, 'AccessControlUnauthorizedAccount');
         });
 
         it('Should not update duration when the voucher type ID is out of bounds', async function () {
-            const { voucherHub, assetEligibilityManager } = await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             await expect(
-                voucherHubWithAssetEligibilityManagerSigner.updateVoucherTypeDuration(
-                    999,
-                    newDuration,
-                ),
+                voucherHubWithManagerSigner.updateVoucherTypeDuration(999, newDuration),
             ).to.be.revertedWith('VoucherHub: type index out of bounds');
         });
     });
 
     describe('Asset Eligibility', function () {
         it('Should set and unset asset eligibility', async function () {
-            const { voucherHub, assetEligibilityManager } = await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             const typeId = await getVoucherTypeCreatedId(voucherHub);
-            const addEligibleAssetTx =
-                await voucherHubWithAssetEligibilityManagerSigner.addEligibleAsset(typeId, asset);
+            const addEligibleAssetTx = await voucherHubWithManagerSigner.addEligibleAsset(
+                typeId,
+                asset,
+            );
             await addEligibleAssetTx.wait();
             expect(addEligibleAssetTx).to.emit(voucherHub, 'EligibleAssetAdded');
             expect(await voucherHub.isAssetEligibleToMatchOrdersSponsoring(typeId, asset)).to.be
                 .true;
-            const removeEligibleAssetTx =
-                await voucherHubWithAssetEligibilityManagerSigner.removeEligibleAsset(
-                    typeId,
-                    asset,
-                );
+            const removeEligibleAssetTx = await voucherHubWithManagerSigner.removeEligibleAsset(
+                typeId,
+                asset,
+            );
             await removeEligibleAssetTx.wait();
             expect(removeEligibleAssetTx).to.emit(voucherHub, 'EligibleAssetRemoved');
             expect(await voucherHub.isAssetEligibleToMatchOrdersSponsoring(typeId, asset)).to.be
@@ -312,27 +254,21 @@ describe('VoucherHub', function () {
         });
 
         it('Should not set asset eligibility when the caller is not authorized', async function () {
-            const { voucherHub, assetEligibilityManager, anyone } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             await expect(
                 voucherHubWithAnyoneSigner.addEligibleAsset(0, asset),
             ).to.be.revertedWithCustomError(voucherHub, 'AccessControlUnauthorizedAccount');
         });
 
         it('Should not unset asset eligibility when the caller is not authorized', async function () {
-            const { voucherHub, assetEligibilityManager, anyone } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             const typeId = await getVoucherTypeCreatedId(voucherHub);
-            const addEligibleAssetTx =
-                await voucherHubWithAssetEligibilityManagerSigner.addEligibleAsset(typeId, asset);
+            const addEligibleAssetTx = await voucherHubWithManagerSigner.addEligibleAsset(
+                typeId,
+                asset,
+            );
             await addEligibleAssetTx.wait();
             expect(await voucherHub.isAssetEligibleToMatchOrdersSponsoring(typeId, asset)).to.be
                 .true;
@@ -344,18 +280,14 @@ describe('VoucherHub', function () {
 
     describe('Create voucher', function () {
         it('Should create and initialize voucher', async function () {
-            const { beacon, voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { beacon, voucherHub, voucherOwner1 } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             const expectedVoucherHubInitRlcBalance = initVoucherHubBalance;
             // Create voucher.
             const voucherHubInitialSrlcBalance = await iexecPocoInstance.balanceOf(
                 voucherHub.getAddress(),
             );
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner
+            const createVoucherTx = await voucherHubWithMinterSigner
                 .createVoucher(voucherOwner1, voucherType, voucherValue)
                 .then((tx) => tx.wait());
 
@@ -408,28 +340,15 @@ describe('VoucherHub', function () {
             // Vouchers are created with the same configuration (type, expiration, ...).
             // The goal is to make sure that configuration is not included in the constructor
             // args which would result in different create2 salts.
-            const {
-                voucherHub,
-                assetEligibilityManager,
-                voucherManager,
-                voucherOwner1,
-                voucherOwner2,
-            } = await loadFixture(deployFixture);
+            const { voucherHub, voucherOwner1, voucherOwner2 } = await loadFixture(deployFixture);
             // Create type0.
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             const voucherHubInitialSrlBalance = await iexecPocoInstance.balanceOf(
                 voucherHub.getAddress(),
             );
             // Create voucher1.
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
-                    voucherOwner1,
-                    voucherType,
-                    voucherValue,
-                ),
+                voucherHubWithMinterSigner.createVoucher(voucherOwner1, voucherType, voucherValue),
             ).to.emit(voucherHub, 'VoucherCreated');
             const voucherHubFirstCreationSrlcBalance = await iexecPocoInstance.balanceOf(
                 voucherHub.getAddress(),
@@ -439,11 +358,7 @@ describe('VoucherHub', function () {
             const voucher1: Voucher = await commonUtils.getVoucher(voucherAddress1);
             // Create voucher2.
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
-                    voucherOwner2,
-                    voucherType,
-                    voucherValue,
-                ),
+                voucherHubWithMinterSigner.createVoucher(voucherOwner2, voucherType, voucherValue),
             ).to.emit(voucherHub, 'VoucherCreated');
             const voucherHubSecondCreationSrlcBalance = await iexecPocoInstance.balanceOf(
                 voucherHub.getAddress(),
@@ -467,21 +382,15 @@ describe('VoucherHub', function () {
         it('Should create multiple vouchers with the correct config', async function () {
             const { voucherHub, voucherOwner1, voucherOwner2 } = await loadFixture(deployFixture);
             // Create type0.
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             const voucherType1 = 1;
             const duration1 = 7200;
             const description1 = 'Long Term Duration';
             const voucherValue1 = 200n;
             // Create type1.
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description1,
-                duration1,
-            );
+            await voucherHubWithManagerSigner.createVoucherType(description1, duration1);
             // Create voucher1.
-            const createVoucherTx1 = await voucherHubWithVoucherManagerSigner.createVoucher(
+            const createVoucherTx1 = await voucherHubWithMinterSigner.createVoucher(
                 voucherOwner1,
                 voucherType,
                 voucherValue,
@@ -495,7 +404,7 @@ describe('VoucherHub', function () {
             const voucher1 = await commonUtils.getVoucher(voucherAddress1);
             const voucherAsProxy1 = await commonUtils.getVoucherAsProxy(voucherAddress1);
             // Create voucher2.
-            const createVoucherTx2 = await voucherHubWithVoucherManagerSigner.createVoucher(
+            const createVoucherTx2 = await voucherHubWithMinterSigner.createVoucher(
                 voucherOwner2,
                 voucherType1,
                 voucherValue1,
@@ -559,59 +468,32 @@ describe('VoucherHub', function () {
         });
 
         it('Should not create more than 1 voucher for the same account', async function () {
-            const { voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub, voucherOwner1 } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             // Create voucher.
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
-                    voucherOwner1,
-                    voucherType,
-                    voucherValue,
-                ),
+                voucherHubWithMinterSigner.createVoucher(voucherOwner1, voucherType, voucherValue),
             ).to.emit(voucherHub, 'VoucherCreated');
             // Second creation should fail.
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
-                    voucherOwner1,
-                    voucherType,
-                    voucherValue,
-                ),
+                voucherHubWithMinterSigner.createVoucher(voucherOwner1, voucherType, voucherValue),
             ).to.be.revertedWithoutReason();
         });
 
         it('Should not create more than 1 voucher for the same account with different config', async function () {
-            const { voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub, voucherOwner1 } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             // Create voucher.
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
-                    voucherOwner1,
-                    voucherType,
-                    voucherValue,
-                ),
+                voucherHubWithMinterSigner.createVoucher(voucherOwner1, voucherType, voucherValue),
             ).to.emit(voucherHub, 'VoucherCreated');
             // Second creation should fail.
             const duration1 = 7200;
             const description1 = 'Long Term Duration';
             const voucherType1 = 1;
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description1,
-                duration1,
-            );
+            await voucherHubWithManagerSigner.createVoucherType(description1, duration1);
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
-                    voucherOwner1,
-                    voucherType1,
-                    voucherValue,
-                ),
+                voucherHubWithMinterSigner.createVoucher(voucherOwner1, voucherType1, voucherValue),
             ).to.be.revertedWithoutReason();
         });
 
@@ -620,14 +502,10 @@ describe('VoucherHub', function () {
         });
 
         it('Should not initialize voucher more than once', async function () {
-            const { voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
-                await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            const { voucherHub, voucherOwner1 } = await loadFixture(deployFixture);
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             // Create voucher.
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner
+            const createVoucherTx = await voucherHubWithMinterSigner
                 .createVoucher(voucherOwner1, voucherType, voucherValue)
                 .then((tx) => tx.wait());
             const expectedExpiration = await commonUtils.getExpectedExpiration(
@@ -657,11 +535,11 @@ describe('VoucherHub', function () {
         });
 
         it('Should not create voucher when voucher type ID is out of bounds', async function () {
-            const { voucherHub, voucherManager, voucherOwner1 } = await loadFixture(deployFixture);
+            const { voucherOwner1 } = await loadFixture(deployFixture);
             const outOfBoundsTypeID = 999;
             // Create voucher.
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
+                voucherHubWithMinterSigner.createVoucher(
                     voucherOwner1,
                     outOfBoundsTypeID,
                     voucherValue,
@@ -671,18 +549,11 @@ describe('VoucherHub', function () {
 
         it('Should not create voucher when SLRC transfer fails', async function () {
             const { voucherOwner1 } = await loadFixture(deployFixture);
-            await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
-                description,
-                duration,
-            );
+            await voucherHubWithManagerSigner.createVoucherType(description, duration);
             await iexecPocoInstance.willFailOnTransfer().then((tx) => tx.wait());
             // Create voucher.
             await expect(
-                voucherHubWithVoucherManagerSigner.createVoucher(
-                    voucherOwner1,
-                    voucherType,
-                    voucherValue,
-                ),
+                voucherHubWithMinterSigner.createVoucher(voucherOwner1, voucherType, voucherValue),
             ).to.be.revertedWith('VoucherHub: SRLC transfer to voucher failed');
         });
     });
@@ -694,10 +565,10 @@ describe('VoucherHub', function () {
 
         beforeEach(async function () {
             ({ voucherHub, voucherOwner1, anyone } = await loadFixture(deployFixture));
-            await voucherHubWithAssetEligibilityManagerSigner
+            await voucherHubWithManagerSigner
                 .createVoucherType(description, duration)
                 .then((tx) => tx.wait());
-            voucherAddress = await voucherHubWithVoucherManagerSigner
+            voucherAddress = await voucherHubWithMinterSigner
                 .createVoucher(voucherOwner1, voucherType, voucherValue)
                 .then((tx) => tx.wait())
                 .then(() => voucherHub.getVoucher(voucherOwner1));
@@ -712,10 +583,7 @@ describe('VoucherHub', function () {
                 anyone,
             ).getExpiration();
 
-            const tx = await voucherHubWithVoucherManagerSigner.topUpVoucher(
-                voucherAddress,
-                topUpValue,
-            );
+            const tx = await voucherHubWithMinterSigner.topUpVoucher(voucherAddress, topUpValue);
             const txReceipt = await tx.wait();
             const expectedExpiration = await commonUtils.getExpectedExpiration(duration, txReceipt);
             await expect(tx)
@@ -742,13 +610,13 @@ describe('VoucherHub', function () {
 
         it('Should not top up voucher without value', async function () {
             await expect(
-                voucherHubWithVoucherManagerSigner.topUpVoucher(Wallet.createRandom().address, 0),
+                voucherHubWithMinterSigner.topUpVoucher(Wallet.createRandom().address, 0),
             ).to.revertedWith('VoucherHub: no value');
         });
 
         it('Should not top up unknown voucher', async function () {
             await expect(
-                voucherHubWithVoucherManagerSigner.topUpVoucher(
+                voucherHubWithMinterSigner.topUpVoucher(
                     Wallet.createRandom().address,
                     voucherValue,
                 ),
@@ -759,7 +627,7 @@ describe('VoucherHub', function () {
             await iexecPocoInstance.willFailOnTransfer().then((tx) => tx.wait());
             // Create voucher.
             await expect(
-                voucherHubWithVoucherManagerSigner.topUpVoucher(voucherAddress, topUpValue),
+                voucherHubWithMinterSigner.topUpVoucher(voucherAddress, topUpValue),
             ).to.be.revertedWith('VoucherHub: SRLC transfer to voucher failed');
         });
     });
@@ -772,15 +640,15 @@ describe('VoucherHub', function () {
             ({ voucherHub, voucherOwner1, voucherOwner2, anyone } =
                 await loadFixture(deployFixture));
             // Create voucher type
-            await voucherHubWithAssetEligibilityManagerSigner
+            await voucherHubWithManagerSigner
                 .createVoucherType(description, duration)
                 .then((tx) => tx.wait());
             // Add eligible asset
-            await voucherHubWithAssetEligibilityManagerSigner
+            await voucherHubWithManagerSigner
                 .addEligibleAsset(voucherType, asset)
                 .then((tx) => tx.wait());
             // Create voucher
-            voucher = await voucherHubWithVoucherManagerSigner
+            voucher = await voucherHubWithMinterSigner
                 .createVoucher(voucherOwner1, voucherType, voucherValue)
                 .then((tx) => tx.wait())
                 .then(() => voucherHub.getVoucher(voucherOwner1))
@@ -848,7 +716,7 @@ describe('VoucherHub', function () {
         });
 
         it('Should debit zero when voucher balance is empty', async function () {
-            const emptyVoucher = await voucherHubWithVoucherManagerSigner
+            const emptyVoucher = await voucherHubWithMinterSigner
                 .createVoucher(voucherOwner2, voucherType, 0)
                 .then((tx) => tx.wait())
                 .then(() => voucherHub.getVoucher(voucherOwner2))
@@ -923,15 +791,15 @@ describe('VoucherHub', function () {
         beforeEach(async function () {
             ({ voucherHub, voucherOwner1, anyone } = await loadFixture(deployFixture));
             // Create voucher type
-            await voucherHubWithAssetEligibilityManagerSigner
+            await voucherHubWithManagerSigner
                 .createVoucherType(description, duration)
                 .then((tx) => tx.wait());
             // Add eligible asset
-            await voucherHubWithAssetEligibilityManagerSigner
+            await voucherHubWithManagerSigner
                 .addEligibleAsset(voucherType, asset)
                 .then((tx) => tx.wait());
             // Create voucher
-            voucher = await voucherHubWithVoucherManagerSigner
+            voucher = await voucherHubWithMinterSigner
                 .createVoucher(voucherOwner1, voucherType, voucherValue)
                 .then((tx) => tx.wait())
                 .then(() => voucherHub.getVoucher(voucherOwner1))
@@ -1002,11 +870,11 @@ describe('VoucherHub', function () {
         beforeEach(async function () {
             ({ voucherHub, voucherOwner1, anyone } = await loadFixture(deployFixture));
             // Create voucher type
-            await voucherHubWithAssetEligibilityManagerSigner
+            await voucherHubWithManagerSigner
                 .createVoucherType(description, duration)
                 .then((tx) => tx.wait());
             // Create voucher
-            voucherAddress = await voucherHubWithVoucherManagerSigner
+            voucherAddress = await voucherHubWithMinterSigner
                 .createVoucher(voucherOwner1, voucherType, voucherValue)
                 .then((tx) => tx.wait())
                 .then(() => voucherHub.getVoucher(voucherOwner1));
@@ -1038,9 +906,7 @@ describe('VoucherHub', function () {
 
         it('Should not drain if unknown voucher', async function () {
             await expect(
-                voucherHubWithVoucherManagerSigner.drainVoucher(
-                    ethers.Wallet.createRandom().address,
-                ),
+                voucherHubWithMinterSigner.drainVoucher(ethers.Wallet.createRandom().address),
             ).to.be.revertedWith('VoucherHub: nothing to drain');
         });
 
@@ -1049,12 +915,10 @@ describe('VoucherHub', function () {
             const expirationDate = await voucher.getExpiration();
             await time.setNextBlockTimestamp(expirationDate); // after expiration
             // Drain to empty the voucher from its balance.
-            await voucherHubWithVoucherManagerSigner
-                .drainVoucher(voucherAddress)
-                .then((tx) => tx.wait());
+            await voucherHubWithMinterSigner.drainVoucher(voucherAddress).then((tx) => tx.wait());
             // Drain a second time when balance is empty.
             await expect(
-                voucherHubWithVoucherManagerSigner.drainVoucher(voucherAddress),
+                voucherHubWithMinterSigner.drainVoucher(voucherAddress),
             ).to.be.revertedWith('VoucherHub: nothing to drain');
         });
     });
@@ -1070,12 +934,7 @@ describe('VoucherHub', function () {
             expect(await iexecPocoInstance.balanceOf(voucherHubAddress)).to.equal(
                 initVoucherHubBalance,
             );
-            await expect(
-                voucherHubWithAssetEligibilityManagerSigner.withdraw(
-                    receiver,
-                    initVoucherHubBalance,
-                ),
-            )
+            await expect(voucherHubWithManagerSigner.withdraw(receiver, initVoucherHubBalance))
                 .to.emit(iexecPocoInstance, 'Transfer')
                 .withArgs(voucherHubAddress, receiver, initVoucherHubBalance);
             expect(await iexecPocoInstance.balanceOf(voucherHubAddress)).to.equal(0);
@@ -1087,7 +946,7 @@ describe('VoucherHub', function () {
                 initVoucherHubBalance,
             );
             const amount = initVoucherHubBalance / 2n;
-            await expect(voucherHubWithAssetEligibilityManagerSigner.withdraw(receiver, amount))
+            await expect(voucherHubWithManagerSigner.withdraw(receiver, amount))
                 .to.emit(iexecPocoInstance, 'Transfer')
                 .withArgs(voucherHubAddress, receiver, amount);
             expect(await iexecPocoInstance.balanceOf(voucherHubAddress)).to.equal(